envs/support_triage_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-triage-env"
version = "0.1.0"
description = "Deterministic customer support triage benchmark for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "support_triage_env.server.app:main"

[tool.setuptools]
packages = ["support_triage_env", "support_triage_env.server"]
package-dir = { "support_triage_env" = ".", "support_triage_env.server" = "server" }

[tool.setuptools.package-data]
support_triage_env = ["**/*.yaml", "**/*.yml"]
